+++ Bencodex.Tests/Types/NullTest.cs
@@ -10,5 +10,17 @@ namespace Bencodex.Tests.Types
         {
             Assert.IsType<Null>(Null.Value);
         }
+
+        [Fact]
+        public void EncodingLength()
+        {
+            Assert.Equal(1, Null.Value.EncodingLength);
+        }
+
+        [Fact]
+        public void Type()
+        {
+            Assert.Equal(ValueType.Null, Null.Value.Type);
+        }
     }
 }