+++ CHANGES.md
@@ -6,6 +6,8 @@ Version 0.4.0
 
 To be released.
 
+ -  Added `Bencodex.Types.ValueType` enum type.
+ -  Added `IValue.Type` property.
  -  Added `IValue.EncodingLength` property.
  -  Removed `Bencodex.Types.CommonVariables` static class.